Create utils methods
